backend/requirements.txt
Flask
Flask-CORS
opencv-python
mediapipe
numpy
Pillow
gunicorn